+++ src/main/java/desbytes/models/Product.java
@@ -5,18 +5,18 @@ package desbytes.models;
  */
 public class Product {
 
-    private int product_id;
+    private double product_id;
     private String name;
     private float price;
 
 
-    public Product(int product_id, String name, float price) {
+    public Product(double product_id, String name, float price) {
         this.product_id = product_id;
         this.name = name;
         this.price = price;
     }
 
-    public int getProduct_id() {
+    public double getProduct_id() {
         return product_id;
     }
